KS-1-4 refactor

+++ app/src/main/kotlin/practice/lesson_1_4 practice.kt
@@ -2,12 +2,12 @@ package practice
 
 fun main () {
 
-    val length: Long = 40868600000
-    val age: Byte = 27
+    val length: ULong = 40868600000u
+    val age: UByte = 27u
     val partOfDay: Float = 0.075f
-    val seconds: Int = 6480
+    val seconds: UShort = 6480u
     val partOfYear: Double = 2.0547945205479453E-4
-    val apogee: Int = 327000
+    val apogee: UInt = 327000u
 
     println(length)
     println(age)
